Unit of work added and Query Optimized

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit() {
     this.loadMessages();
@@ -31,9 +32,13 @@ export class MessagesComponent implements OnInit {
     })
   }
 
-  deleteMessage(id: number){
-    this.messageService.deleteMessage(id).subscribe(response => {
-      this.messages.splice(this.messages.findIndex(i => i.id === id), 1);
+  deleteMessage(id: number) {
+    this.confirmService.confirm('Confirm Message', 'If you delete the message, it cannot be undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(response => {
+          this.messages.splice(this.messages.findIndex(i => i.id === id), 1);
+        })
+      }
     })
   }
